Mix v1.11.0-dev Mix.Shell.Quiet

This is Mix's default shell when the MIX_QUIET environment variable is set.

It's just like Mix.Shell.IO, but prints far less.

Functions

cmd(command, opts \\ [])

Executes the given command quietly without outputting anything.

error(message)

Prints the error to the shell followed by a newline.

info(message)

Prints nothing to the shell.

print_app()

Prints the current application if it was not printed yet.

prompt(message)

Prints a message and prompts the user for input.

Input will be consumed until Enter is pressed.

yes?(message)

Prints a message and asks the user if they want to proceed.

The user must press Enter or type one of "y", "yes", "Y", "YES" or "Yes".