Mix v1.10.3 Mix.Task behaviour

A simple module that provides conveniences for creating, loading and manipulating tasks.

A Mix task can be defined by simply using Mix.Task in a module starting with Mix.Tasks. and defining the run/1 function:

defmodule Mix.Tasks.Echo do
  use Mix.Task

  @impl Mix.Task
  def run(args) do
    Mix.shell().info(Enum.join(args, " "))
  end
end

The run/1 function will receive a list of all arguments passed to the command line.

Attributes

There are a few attributes available in Mix tasks to configure them in Mix:

@shortdoc - makes the task public with a short description that appears on mix help

@recursive - runs the task recursively in umbrella projects

@preferred_cli_env - recommends environment to run task. It is used in absence of a Mix project recommendation, or explicit MIX_ENV, and it only works for tasks in the current project. @preferred_cli_env is not loaded from dependencies as we need to know the environment before dependencies are loaded.

Documentation

Users can read the documentation for public Mix tasks by running mix help my_task. The documentation that will be shown is the @moduledoc of the task's module.

Types

task_module()

Specs

task_module() :: atom()

task_name()

Specs

task_name() :: String.t() | atom()

Functions

alias?(task)

Specs

alias?(task_name()) :: boolean()

Checks if an alias called task exists.

For more information about task aliasing, take a look at the "Aliasing" section in the docs for Mix.

all_modules()

Specs

all_modules() :: [task_module()]

Returns all loaded task modules.

Modules that are not yet loaded won't show up. Check load_all/0 if you want to preload all tasks.

clear()

Specs

clear() :: :ok

Clears all invoked tasks, allowing them to be reinvoked.

This operation is not recursive.

get(task)

Specs

get(task_name()) :: task_module() | nil

Receives a task name and returns the task module if found.

Otherwise returns nil in case the module exists, but it isn't a task or cannot be found.

get!(task)

Specs

get!(task_name()) :: task_module()

Receives a task name and retrieves the task module.

Exceptions

Mix.NoTaskError - raised if the task could not be found

Mix.InvalidTaskError - raised if the task is not a valid Mix.Task

load_all()

Specs

load_all() :: [task_module()]

Loads all tasks in all code paths.

load_tasks(dirs)

Specs

load_tasks([List.Chars.t()]) :: [task_module()]

Loads all tasks in the given paths.

moduledoc(module)

Specs

moduledoc(task_module()) :: String.t() | nil | false

Gets the moduledoc for the given task module.

Returns the moduledoc or nil.

preferred_cli_env(task)

Specs

preferred_cli_env(task_name()) :: atom() | nil

Gets preferred CLI environment for the task.

Returns environment (for example, :test, or :prod), or nil.

recursing?()
(since 1.8.0)

Specs

recursing?() :: boolean()

Indicates if the current task is recursing.

This returns true if a task is marked as recursive and it is being executed inside an umbrella project.

recursive(module)

Specs

recursive(task_module()) :: boolean()

Checks if the task should be run recursively for all sub-apps in umbrella projects.

Returns true or false.

reenable(task)

Specs

reenable(task_name()) :: :ok

Reenables a given task so it can be executed again down the stack.

Both alias and the regular stack are reenabled when this function is called.

If an umbrella project reenables a task, it is reenabled for all child projects.

rerun(task, args \\ [])

Specs

rerun(task_name(), [any()]) :: any()

Reruns task with the given arguments.

This function reruns the given task; to do that, it first re-enables the task and then runs it as normal.

run(task, args \\ [])

Specs

run(task_name(), [any()]) :: any()

Runs a task with the given args.

If the task was not yet invoked, it runs the task and returns the result.

If there is an alias with the same name, the alias will be invoked instead of the original task.

If the task or alias were already invoked, it does not run them again and simply aborts with :noop.

It may raise an exception if an alias or a task can't be found or the task is invalid. Check get!/1 for more information.

shortdoc(module)

Specs

shortdoc(task_module()) :: String.t() | nil

Gets the shortdoc for the given task module.

Returns the shortdoc or nil.

task?(module)

Specs

task?(task_module()) :: boolean()

Returns true if given module is a task.

task_name(module)

Specs

task_name(task_module()) :: task_name()

Returns the task name for the given module.

Callbacks

run(command_line_args)

Specs

run(command_line_args :: [binary()]) :: any()

A task needs to implement run which receives a list of command line args.